Moving out sshmanager from patroni, renaming pgconnector to pgmanager
