+++ pgmanager.go
